backend/requirements.txt
fastapi == 0.65.1

uvicorn == 0.14.0

motor == 3.1.2

python-telegram-bot
